Fix quantizer hysteresis bug
The bug can be demonstrated by calling quantizeValue() with an argument
alternating between two adjacent values greater than 255 which result
in different return values. One such pair of values is 301 and 302. The
quantizer will update every time even though it should only update when
the input changes by more than 2 LSBs.

+++ Firmware/quantizer.c
@@ -115,7 +115,7 @@ ISR(PCINT2_vect)
     return;	
 };
 //-----------------------------------------------------------
-static uint8_t lastInput=0;
+static uint16_t lastInput=0;
 uint8_t quantizeValue(uint16_t input)
 {
   if(io_getActiveSteps()==0)
